# frontend/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/login":{"post":{"tags":["controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/user":{"get":{"tags":["controller"],"operationId":"getUser","parameters":[{"name":"session","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/user/sync":{"get":{"tags":["controller"],"operationId":"syncUser","parameters":[{"name":"session","in":"query","required":true,"schema":{"type":"string"}},{"name":"portalId","in":"query","required":true,"schema":{"type":"string"}},{"name":"portalPassword","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/user/recommended-course":{"get":{"tags":["controller"],"operationId":"recommendCourse","parameters":[{"name":"session","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Course"}}}}}}}},"/api/user/full":{"get":{"tags":["controller"],"operationId":"getUserFullInfo","parameters":[{"name":"session","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserFullInfo"}}}}}}},"/api/user/academic-status":{"get":{"tags":["controller"],"operationId":"getUserAcademicStatus","parameters":[{"name":"session","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserAcademicStatusDTO"}}}}}}},"/api/logout":{"get":{"tags":["controller"],"operationId":"logout","parameters":[{"name":"session","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/check":{"get":{"tags":["controller"],"operationId":"check","parameters":[{"name":"session","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/":{"get":{"tags":["controller"],"operationId":"index","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"LoginDTO":{"type":"object","properties":{"portalId":{"type":"string"},"portalPassword":{"type":"string"}}},"User":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"studentId":{"type":"string"},"department":{"type":"string","enum":["ComputerScience","GeneralEducation"]},"currentGrade":{"type":"integer","format":"int32"},"currentSemester":{"type":"integer","format":"int32"},"portalId":{"type":"string"},"hashedPortalPassword":{"type":"string"},"salt":{"type":"string"},"createdAt":{"type":"integer","format":"int64"},"updatedAt":{"type":"integer","format":"int64"},"password":{"type":"string","writeOnly":true}}},"Course":{"type":"object","properties":{"id":{"type":"string"},"lectureName":{"type":"string"},"lectureCode":{"type":"string"},"lectureCredit":{"type":"integer","format":"int32"},"lectureGrade":{"type":"integer","format":"int32"},"generalCredit":{"type":"integer","format":"int32"},"designCredit":{"type":"integer","format":"int32"},"majorElective":{"type":"boolean"},"majorEssential":{"type":"boolean"},"liberalElective":{"type":"boolean"},"liberalEssential":{"type":"boolean"},"engineering":{"type":"boolean"},"basicAcademic":{"type":"boolean"}}},"UserFullInfo":{"type":"object","properties":{"name":{"type":"string"},"studentId":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"major":{"type":"string"},"grade":{"type":"integer","format":"int32"},"status":{"type":"string"},"courses":{"type":"array","items":{"$ref":"#/components/schemas/Course"}},"updatedAt":{"type":"integer","format":"int64"}}},"UserAcademicStatusDTO":{"type":"object","properties":{"totalRequiredCredit":{"type":"integer","format":"int32"},"totalCompletedCredit":{"type":"integer","format":"int32"},"majorRequiredCredit":{"type":"integer","format":"int32"},"majorCompletedCredit":{"type":"integer","format":"int32"},"majorEssentialRequiredCredit":{"type":"integer","format":"int32"},"majorEssentialCompletedCredit":{"type":"integer","format":"int32"},"liberalRequiredCredit":{"type":"integer","format":"int32"},"liberalCompletedCredit":{"type":"integer","format":"int32"},"liberalEssentialRequiredCredit":{"type":"integer","format":"int32"},"liberalEssentialCompletedCredit":{"type":"integer","format":"int32"},"engineeringRequiredCredit":{"type":"integer","format":"int32"},"engineeringCompletedCredit":{"type":"integer","format":"int32"},"generalRequiredCredit":{"type":"integer","format":"int32"},"generalCompletedCredit":{"type":"integer","format":"int32"},"minimumTotalGradePointAverage":{"type":"number","format":"double"},"totalGradePointAverage":{"type":"number","format":"double"}}}}}}